requirements.txt
RPi.GPIO==0.7.0
dot3k==2.0.1
redis==3.5.3